Sonar sweep: ~90 of 114 open issues fixed, two real bugs among them
Seven parallel agents over disjoint file sets, zero-behavior-change rules,
then one integration pass. The mechanical bulk: duplicated literals hoisted
to constants, redundant casts dropped (Math.floorMod(long,int) already
returns int), unused code removed, dangling Javadoc reattached, restricted
identifiers renamed (ReputationService.record -> recordCrime), and twenty
cognitive-complexity refactors done as pure method extraction - worldgen
extractions preserve Random draw order, so seeded worlds are byte-identical.

The two real bugs (S2184): int arithmetic widened too late in
GalaxyEngine.rollBand and StarChartRenderer's text positioning.

The find of the day: Settings.java can NEVER hold static constants -
BentoBox's YAML loader PropertyDescriptors every declared field and throws
on any without a getter/setter, so the S1192 fix there booted the addon
DISABLED with null settings. TradeWindsTest caught it (the only test that
loads Settings through the real YAML path). Reverted to literals; the trap
is now documented in the class and in CLAUDE.md.

Deliberately not fixed (mark won't-fix on SonarCloud): gameplay-random
S2245s, the /twadmin warpfail "debug feature", the GameModeAddon wiring
hub S6539, EncounterType's data-table enum ctor, BentoBox-compat
deprecations, and Settings' literals per the loader trap.

+++ src/main/java/world/bentobox/tradewinds/commands/TWSpawnCommand.java
@@ -119,7 +119,7 @@ public class TWSpawnCommand extends DelayedTeleportCommand {
         String[] points = { "north", "north-east", "east", "south-east", "south", "south-west", "west",
                 "north-west" };
         double angle = Math.toDegrees(Math.atan2(dx, -dz));
-        int index = (int) Math.floorMod(Math.round(angle / 45.0), 8);
+        int index = Math.floorMod(Math.round(angle / 45.0), 8);
         return "tradewinds.direction." + points[index];
     }
 }
